[ADD] add image average energy extractor

+++ calculate_energy.py
@@ -12,7 +12,7 @@ tifs = [x for x in os.listdir(raw_dir) if 'tif' in x.lower()]
 
 for tif in tifs:
     fpath = os.path.join(raw_dir,tif)
-    img = cv2.imread(fpath)
+    img = cv2.imread(fpath).astype(np.float)
     energy = float(np.sqrt(np.mean(np.square(img))))
     energy_dict[tif.split('.')[0]] = energy
